#84 Support for language id and full file name extensions
Some of the language server implementation extension and language id are
different. So then the specific language id must be provided. Now the
RawCommandServerDefinition.scala support excepting language id as well.

Some of the language server implementation are based on cetain file
names than the extension. Now as a fallback option if the extension
(which is more prominent approach) is not found the LSP will look for
file name mapping as well.

Signed-off-by: gayanper <gayanper@gmail.com>